Espaçamento entre os tiles

+++ core/src/com/devmobile/game/helpers/GameInfo.java
@@ -3,7 +3,7 @@ package com.devmobile.game.helpers;
 public class GameInfo {
     public final static float WIDHT = 670;
     public final static float HEIGHT = 375;
-    public final static float sizeTexture = 16;
+    public final static int sizeTexture = 16;
     public final static int maxY = Math.round(GameInfo.HEIGHT / GameInfo.sizeTexture);
     public static String grasslandTerrain = "Grassland_Terrain_";
     public final static String autumnForestTerrain = "AutumnForest_Terrain_";
